feat: ProfileCardList에 인터렉티브 애니메이션 추가
