+++ src/shared/ui/profile/ui/profileCardList/ProfileCardList.styles.ts
@@ -1,0 +1,16 @@
+import styled from "@emotion/styled";
+
+export const ProfileCardListContainer = styled.div`
+  width: 100%;
+  height: 100%;
+  .swiper-slide-prev{
+    opacity: 0.8;
+    filter: blur(2px);
+    transition: opacity 0.3s ease-in-out, filter 0.3s ease-in-out;
+  }
+  .swiper-slide-next{
+    opacity: 0.8;
+    filter: blur(2px);
+    transition: opacity 0.3s ease-in-out, filter 0.3s ease-in-out;
+  }
+`;
