+++ src/shared/ui/project/ui/projectCardList/ProjectCardList.tsx
@@ -3,9 +3,6 @@ import ProjectCard from '../projectCard/ProjectCard';
 import * as S from './ProjectCardList.styles';
 import { Swiper, SwiperSlide } from 'swiper/react';
 import { Navigation, Pagination } from 'swiper/modules';
-import 'swiper/css';
-import 'swiper/css/navigation';
-import 'swiper/css/pagination';
 
 interface ProjectCardListProps {
   projects: ProjectDataType[];
